requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sse-starlette==2.1.0
langchain==0.3.25
langchain-google-genai==2.1.4
langgraph==0.4.1
chromadb==0.6.3
sqlalchemy==2.0.35
python-dotenv==1.0.1
pydantic==2.9.2
httpx==0.27.2
